Composer Update Add conditon to check if image exits

+++ app/Http/Controllers/PostsController.php
@@ -108,7 +108,10 @@ class PostsController extends Controller
         $post = Post::withTrashed()->where('id', $id)->firstOrFail();
 
         if($post->trashed()){
-            Storage::delete($post->image);
+            if(Storage::exists($post->image))
+            {
+                Storage::delete($post->image);
+            }
             $post->forceDelete();
         } else {
             $post->delete();
